Capture snapshot stream in anticipation of displaying thumbnails
This patch set sets the stage for displaying frame thumbnails in
qapitrace by first capturing a snaphot stream.

First we set prepare qapitrace to capture a snapshot stream. We enhance
replayTrace() to accept a second parameter for specifying snapshot
dumping. This then enables us to notify the retracer so set up command
line parameters to glretrace, which runs out of process. We specify that
glretrace must generate snapshots and then emit them to stdout.

After glretrace returns, we extract PNM images from the output stream.
We enhance PNM processing to parse the ASCII header to extract number
channels, the width and the height of the following binary image data.
We then extract the binary data of the image on a per-scanline copy. We
create a QImage of the correct size and insert the image data into the
QImage.

Finally, we create a QList collection to hold the sequence of images.
This list is return to the main window object of qapitrace and will
later be used to display those snapshot images as thumbnails.