Display captured snapshots as thumbnails.
Once snapshots are captured after being generated by glretrace, they are
then displayed in the GUI app, qapitrace.

The retracer passes a list of images to the main window. The main window
in turn binds those images to the ApiTraceFrames for later display.

When the ApiTrace events are later displayed, when a frame is being
display, its snapshot image is queried. If an image is available, it is
reduced to create a thumbnail. That thumbnail is then displayed at the
front of the line for a frame.

One bug still exists. The thumbnail only gets displayed when the user
interacts with the window.